requirements.txt
beautifulsoup4==4.9.3
DateTime==4.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fuckit==4.8.1
gunicorn==20.0.4
ipywidgets==7.6.3
matplotlib==3.3.4
pandas-datareader==0.9.0
git://github.com/briantacderan/hoodflex.git#egg=hoodflex
